% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{acset_equal}
\alias{acset_equal}
\title{Test structural equality of two ACSets}
\usage{
acset_equal(x, y)
}
\arguments{
\item{x}{An ACSet.}

\item{y}{An ACSet.}
}
\value{
Logical: \code{TRUE} if the two ACSets are structurally equal.
}
\description{
Test structural equality of two ACSets
}
