% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsets-package.R
\docType{package}
\name{acsets-package}
\alias{acsets}
\alias{acsets-package}
\title{acsets: Attributed C-Sets for R}
\description{
Attributed C-sets (ACSets) are a family of data structures generalizing both graphs and data frames. They are an efficient in-memory implementation of a category-theoretic formalism for relational databases. This package provides data structures for ACSet schemas and ACSets, with operations for querying, mutation, and serialization. Port of the AlgebraicJulia ACSets.jl package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/catRgory/acsets}
  \item \url{https://catrgory.github.io/acsets/}
  \item Report bugs at \url{https://github.com/catRgory/acsets/issues}
}

}
\author{
\strong{Maintainer}: Simon Frost \email{sdwfrost@gmail.com} (\href{https://orcid.org/0000-0002-5207-9879}{ORCID})

}
\keyword{internal}
