% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{add_part}
\alias{add_part}
\title{Add a single part, optionally setting subparts}
\usage{
add_part(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Named subpart values to set on the new part.}
}
\value{
Integer ID of the newly added part.
}
\description{
Add a single part, optionally setting subparts
}
\examples{
sch <- BasicSchema(
  obs = c("E", "V"),
  homs = list(hom("src", "E", "V"), hom("tgt", "E", "V"))
)
g <- ACSet(sch)
add_part(g, "V")
add_part(g, "V")
add_part(g, "E", src = 1L, tgt = 2L)
subpart(g, 1L, "src")
}
