% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{arrows}
\alias{arrows}
\title{Get all arrows (morphisms and attributes) from a schema}
\usage{
arrows(x, ...)
}
\arguments{
\item{x}{A BasicSchema.}

\item{...}{Arguments passed to methods.}
}
\value{
List of arrow specs.
}
\description{
Get all arrows (morphisms and attributes) from a schema
}
