% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Convert ACSet object type to data.frame}
\usage{
as_data_frame(acs, ob)
}
\arguments{
\item{acs}{An ACSet.}

\item{ob}{Character: object type name.}
}
\value{
A data frame with part IDs and subpart columns.
}
\description{
Convert ACSet object type to data.frame
}
