% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{attr_spec}
\alias{attr_spec}
\title{Create an attribute specification}
\usage{
attr_spec(name, dom, codom)
}
\arguments{
\item{name}{Name of the attribute.}

\item{dom}{Domain object name.}

\item{codom}{Codomain attribute type name.}
}
\value{
A named list with elements \code{name}, \code{dom}, \code{codom}.
}
\description{
Create an attribute specification
}
