% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{attrs}
\alias{attrs}
\title{Get attribute specs from a schema}
\usage{
attrs(x, ...)
}
\arguments{
\item{x}{A BasicSchema.}

\item{...}{Arguments passed to methods.}
}
\value{
List of attribute specs.
}
\description{
Get attribute specs from a schema
}
