% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{attrtypes}
\alias{attrtypes}
\title{Get attribute type names from a schema}
\usage{
attrtypes(x, ...)
}
\arguments{
\item{x}{A BasicSchema.}

\item{...}{Arguments passed to methods.}
}
\value{
Character vector of attribute type names.
}
\description{
Get attribute type names from a schema
}
