% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion.R
\name{cascading_rem_part}
\alias{cascading_rem_part}
\title{Remove a part and cascade to dependent parts}
\usage{
cascading_rem_part(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
The ACSet, invisibly.
}
\description{
Remove a part and cascade to dependent parts
}
\examples{
sch <- BasicSchema(
  obs = c("E", "V"),
  homs = list(hom("src", "E", "V"), hom("tgt", "E", "V"))
)
g <- ACSet(sch, index = c("src", "tgt"))
add_parts(g, "V", 3)
add_parts(g, "E", 2, src = c(1L, 2L), tgt = c(2L, 3L))
# Removing vertex 2 also removes edges that reference it
cascading_rem_part(g, "V", 2L)
nparts(g, "E")
}
