% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion.R
\name{cascading_rem_parts}
\alias{cascading_rem_parts}
\title{Remove multiple parts with cascading}
\usage{
cascading_rem_parts(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
The ACSet, invisibly.
}
\description{
Remove multiple parts with cascading
}
