% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{clear_subpart}
\alias{clear_subpart}
\title{Clear a subpart value}
\usage{
clear_subpart(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
The ACSet, invisibly.
}
\description{
Clear a subpart value
}
