% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{codom}
\alias{codom}
\title{Get the codomain of an arrow}
\usage{
codom(x, ...)
}
\arguments{
\item{x}{A BasicSchema.}

\item{...}{Arguments passed to methods.}
}
\value{
Character string naming the codomain.
}
\description{
Get the codomain of an arrow
}
