% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{copy_acset}
\alias{copy_acset}
\title{Create a deep copy of an ACSet}
\usage{
copy_acset(x)
}
\arguments{
\item{x}{An ACSet.}
}
\value{
A new ACSet that is a deep copy of \code{x}.
}
\description{
Create a deep copy of an ACSet
}
