% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{disjoint_union}
\alias{disjoint_union}
\title{Disjoint union of two ACSets with the same schema}
\usage{
disjoint_union(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
A new ACSet containing the disjoint union.
}
\description{
Disjoint union of two ACSets with the same schema
}
