% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{dom}
\alias{dom}
\title{Get the domain of an arrow}
\usage{
dom(x, ...)
}
\arguments{
\item{x}{A BasicSchema.}

\item{...}{Arguments passed to methods.}
}
\value{
Character string naming the domain object.
}
\description{
Get the domain of an arrow
}
