% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{From}
\alias{From}
\title{Start a query on an ACSet object type}
\usage{
From(acs, ob)
}
\arguments{
\item{acs}{An ACSet.}

\item{ob}{Character: object type to query.}
}
\value{
An \code{acset_query} object for piping to \code{\link[=Where]{Where()}} and \code{\link[=Select]{Select()}}.
}
\description{
Start a query on an ACSet object type
}
