% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion.R
\name{gc_acset}
\alias{gc_acset}
\title{Garbage collection for BitSetParts (no-op for IntParts)}
\usage{
gc_acset(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
The ACSet, invisibly.
}
\description{
Garbage collection for BitSetParts (no-op for IntParts)
}
