% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{generate_json_schema}
\alias{generate_json_schema}
\title{Generate a JSON-compatible list from a schema}
\usage{
generate_json_schema(schema)
}
\arguments{
\item{schema}{A BasicSchema.}
}
\value{
A list suitable for JSON serialization.
}
\description{
Generate a JSON-compatible list from a schema
}
