% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{has_part}
\alias{has_part}
\title{Check if a part exists}
\usage{
has_part(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
Logical.
}
\description{
Check if a part exists
}
