% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{has_subpart}
\alias{has_subpart}
\title{Check if a subpart (morphism/attribute) exists in the schema}
\usage{
has_subpart(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
Logical.
}
\description{
Check if a subpart (morphism/attribute) exists in the schema
}
