% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{hom}
\alias{hom}
\title{Create a morphism (foreign key) specification}
\usage{
hom(name, dom, codom)
}
\arguments{
\item{name}{Name of the morphism.}

\item{dom}{Domain object name.}

\item{codom}{Codomain object name.}
}
\value{
A named list with elements \code{name}, \code{dom}, \code{codom}.
}
\description{
Create a morphism (foreign key) specification
}
