% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{homs}
\alias{homs}
\title{Get morphisms (foreign keys) from a schema}
\usage{
homs(x, ...)
}
\arguments{
\item{x}{A BasicSchema.}

\item{...}{Arguments passed to methods.}
}
\value{
List of morphism specs.
}
\description{
Get morphisms (foreign keys) from a schema
}
