% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{incident}
\alias{incident}
\title{Incident query: find parts whose subpart equals a given value}
\usage{
incident(x, ...)
}
\arguments{
\item{x}{An ACSet}

\item{...}{Arguments passed to methods.}
}
\description{
Incident query: find parts whose subpart equals a given value
}
\examples{
sch <- BasicSchema(
  obs = c("E", "V"),
  homs = list(hom("src", "E", "V"), hom("tgt", "E", "V"))
)
g <- ACSet(sch, index = c("src", "tgt"))
add_parts(g, "V", 3)
add_parts(g, "E", 2, src = c(1L, 1L), tgt = c(2L, 3L))
# Which edges have source vertex 1?
incident(g, 1L, "src")
}
