% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{maxpart}
\alias{maxpart}
\title{Maximum part ID for an object type}
\usage{
maxpart(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
Integer maximum part ID.
}
\description{
Maximum part ID for an object type
}
