% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{nparts}
\alias{nparts}
\title{Number of parts of a given object type}
\usage{
nparts(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
Integer count of parts.
}
\description{
Number of parts of a given object type
}
