% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{objects}
\alias{objects}
\title{Get object names from a schema}
\usage{
objects(x, ...)
}
\arguments{
\item{x}{A BasicSchema.}

\item{...}{Arguments passed to methods.}
}
\value{
Character vector of object names.
}
\description{
Get object names from a schema
}
