% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{parse_json_acset}
\alias{parse_json_acset}
\title{Parse a JSON-compatible list into an ACSet}
\usage{
parse_json_acset(constructor, input)
}
\arguments{
\item{constructor}{An acset_constructor from \code{\link[=acset_type]{acset_type()}}.}

\item{input}{A nested list (typically from JSON) representing ACSet data.}
}
\value{
An ACSet instance.
}
\description{
Parse a JSON-compatible list into an ACSet
}
\examples{
sch <- BasicSchema(obs = c("V"), attrtypes = c("Name"),
                   attrs = list(attr_spec("name", "V", "Name")))
Verts <- acset_type(sch)
input <- list(V = list(list(`_id` = 1L, name = "a"),
                       list(`_id` = 2L, name = "b")))
g <- parse_json_acset(Verts, input)
nparts(g, "V")
}
