% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{parse_json_schema}
\alias{parse_json_schema}
\title{Parse a JSON-compatible list into a BasicSchema}
\usage{
parse_json_schema(input)
}
\arguments{
\item{input}{A list (typically from JSON) representing a schema.}
}
\value{
A BasicSchema object.
}
\description{
Parse a JSON-compatible list into a BasicSchema
}
