% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{parts}
\alias{parts}
\title{Get all part IDs for an object type}
\usage{
parts(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
Integer vector of part IDs.
}
\description{
Get all part IDs for an object type
}
