% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{read_json_acset}
\alias{read_json_acset}
\title{Read an ACSet from JSON file}
\usage{
read_json_acset(constructor, path)
}
\arguments{
\item{constructor}{An acset_constructor from \code{\link[=acset_type]{acset_type()}}.}

\item{path}{File path to read from.}
}
\value{
An ACSet instance.
}
\description{
Read an ACSet from JSON file
}
