% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion.R
\name{rem_part}
\alias{rem_part}
\title{Remove a single part using pop-and-swap strategy}
\usage{
rem_part(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
The ACSet, invisibly.
}
\description{
Remove a single part using pop-and-swap strategy
}
