% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion.R
\name{rem_parts}
\alias{rem_parts}
\title{Remove multiple parts (must be processed from largest to smallest)}
\usage{
rem_parts(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
The ACSet, invisibly.
}
\description{
Remove multiple parts (must be processed from largest to smallest)
}
