% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{set_subpart}
\alias{set_subpart}
\title{Set a single subpart value}
\usage{
set_subpart(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
The ACSet, invisibly.
}
\description{
Set a single subpart value
}
