% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{set_subparts}
\alias{set_subparts}
\title{Set multiple subparts at once}
\usage{
set_subparts(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
The ACSet, invisibly.
}
\description{
Set multiple subparts at once
}
