% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tables}
\alias{tables}
\title{Get all tables from an ACSet}
\usage{
tables(acs)
}
\arguments{
\item{acs}{An ACSet.}
}
\value{
A named list of data frames, one per object type.
}
\description{
Get all tables from an ACSet
}
