% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{types}
\alias{types}
\title{Get all type names (objects and attribute types)}
\usage{
types(x, ...)
}
\arguments{
\item{x}{A BasicSchema.}

\item{...}{Arguments passed to methods.}
}
\value{
Character vector of all type names.
}
\description{
Get all type names (objects and attribute types)
}
