% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{write_json_acset}
\alias{write_json_acset}
\title{Write an ACSet to JSON file}
\usage{
write_json_acset(x, path)
}
\arguments{
\item{x}{An ACSet.}

\item{path}{File path to write to.}
}
\value{
The file path, invisibly.
}
\description{
Write an ACSet to JSON file
}
