% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{acset}
\alias{acset}
\title{Construct an ACSet using named arguments or NSE}
\usage{
acset(type, ...)
}
\arguments{
\item{type}{An acset_constructor (from acset_type()) or an ACSet class}

\item{...}{Named arguments: object names get integer counts,
morphism/attribute names get value vectors}
}
\value{
An ACSet instance
}
\description{
Construct an ACSet using named arguments or NSE
}
\examples{
sch <- BasicSchema(
  obs = c("E", "V"),
  homs = list(hom("src", "E", "V"), hom("tgt", "E", "V"))
)
Graph <- acset_type(sch)
g <- acset(Graph, V = 2, E = 1, src = 1L, tgt = 2L)
nparts(g, "E")
}
