% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{acset_type}
\alias{acset_type}
\title{Create an ACSet type bound to a specific schema}
\usage{
acset_type(schema, name = NULL, index = character())
}
\arguments{
\item{schema}{A BasicSchema}

\item{name}{Optional name for the type}

\item{index}{Character vector of morphism/attribute names to index}
}
\value{
A function that creates ACSet instances with this schema
}
\description{
Returns a constructor function that creates ACSets with the given schema.
This is the R equivalent of Julia's \verb{@acset_type}.
}
\examples{
sch <- BasicSchema(
  obs = c("E", "V"),
  homs = list(hom("src", "E", "V"), hom("tgt", "E", "V")),
  attrtypes = c("Name"),
  attrs = list(attr_spec("name", "V", "Name"))
)
Graph <- acset_type(sch, index = c("src", "tgt"))

# Create an empty graph
g <- Graph()

# Create a pre-populated graph
g2 <- Graph(V = 3, E = 2, src = c(1L, 2L), tgt = c(2L, 3L),
            name = c("a", "b", "c"))
nparts(g2, "V")
}
