% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{add_part}
\alias{add_part}
\title{Add a single part, optionally setting subparts}
\usage{
add_part(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
Integer ID of the newly added part.
}
\description{
Add a single part, optionally setting subparts
}
