% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{add_parts}
\alias{add_parts}
\title{Add multiple parts at once}
\usage{
add_parts(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
Integer vector of newly added part IDs.
}
\description{
Add multiple parts at once
}
