% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{add_parts}
\alias{add_parts}
\title{Add multiple parts at once}
\usage{
add_parts(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Named subpart values. Vectors must have length 1, length \code{n}, or
recycle evenly across \code{n}.}
}
\value{
Integer vector of newly added part IDs.
}
\description{
Add multiple parts at once
}
\examples{
sch <- BasicSchema(
  obs = c("V"),
  attrtypes = c("Name"),
  attrs = list(attr_spec("name", "V", "Name"))
)
g <- ACSet(sch)
ids <- add_parts(g, "V", 3, name = c("a", "b", "c"))
ids
subpart(g, NULL, "name")
}
