% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion.R
\name{cascading_rem_part}
\alias{cascading_rem_part}
\title{Remove a part and cascade to dependent parts}
\usage{
cascading_rem_part(x, ...)
}
\arguments{
\item{x}{An ACSet.}

\item{...}{Arguments passed to methods.}
}
\value{
The ACSet, invisibly.
}
\description{
Remove a part and cascade to dependent parts
}
