% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{From}
\alias{From}
\title{Start a query on an ACSet object type}
\usage{
From(acs, ob)
}
\arguments{
\item{acs}{An ACSet.}

\item{ob}{Character: object type to query.}
}
\value{
An \code{acset_query} object for piping to \code{\link[=Where]{Where()}} and \code{\link[=Select]{Select()}}.
}
\description{
Start a query on an ACSet object type
}
\examples{
sch <- BasicSchema(
  obs = c("V"),
  attrtypes = c("Name", "Age"),
  attrs = list(attr_spec("name", "V", "Name"),
               attr_spec("age", "V", "Age"))
)
g <- ACSet(sch)
add_parts(g, "V", 3, name = c("Alice", "Bob", "Carol"),
          age = c(30, 25, 35))
# Query pipeline: find vertices with age > 28
result <- From(g, "V") |> Where("age", `>`, 28) |> Select("name", "age")
result
}
