% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{generate_json_acset}
\alias{generate_json_acset}
\title{Generate JSON-compatible list from an ACSet}
\usage{
generate_json_acset(x)
}
\arguments{
\item{x}{An ACSet.}
}
\value{
A nested list suitable for JSON serialization.
}
\description{
Generate JSON-compatible list from an ACSet
}
