% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acset.R
\name{incident}
\alias{incident}
\title{Incident query: find parts whose subpart equals a given value}
\usage{
incident(x, ...)
}
\arguments{
\item{x}{An ACSet}

\item{...}{Arguments passed to methods.}
}
\description{
Incident query: find parts whose subpart equals a given value
}
