% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{parse_json_acset}
\alias{parse_json_acset}
\title{Parse a JSON-compatible list into an ACSet}
\usage{
parse_json_acset(constructor, input)
}
\arguments{
\item{constructor}{An acset_constructor from \code{\link[=acset_type]{acset_type()}}.}

\item{input}{A nested list (typically from JSON) representing ACSet data.}
}
\value{
An ACSet instance.
}
\description{
Parse a JSON-compatible list into an ACSet
}
