% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Where}
\alias{Where}
\title{Filter query results by a subpart value}
\usage{
Where(query, f, op, value)
}
\arguments{
\item{query}{An \code{acset_query} object from \code{\link[=From]{From()}}.}

\item{f}{Character: morphism or attribute name.}

\item{op}{A comparison function (e.g. \code{==}, \code{<}).}

\item{value}{The value to compare against.}
}
\value{
The filtered \code{acset_query} object.
}
\description{
Filter query results by a subpart value
}
