% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{write_json_acset}
\alias{write_json_acset}
\title{Write an ACSet to JSON file}
\usage{
write_json_acset(x, path)
}
\arguments{
\item{x}{An ACSet.}

\item{path}{File path to write to.}
}
\value{
The file path, invisibly.
}
\description{
Write an ACSet to JSON file
}
\examples{
sch <- BasicSchema(obs = c("V"), attrtypes = c("Name"),
                   attrs = list(attr_spec("name", "V", "Name")))
Verts <- acset_type(sch)
g <- Verts(V = 2, name = c("a", "b"))
tmp <- tempfile(fileext = ".json")
write_json_acset(g, tmp)
g2 <- read_json_acset(Verts, tmp)
acset_equal(g, g2)
unlink(tmp)
}
